requirements.txt
playwright==1.42.0
requests==2.31.0
pygame==2.5.2
python-dotenv==1.0.1
groq==0.4.2
ipdb==0.13.13
